+++ src/commonMain/kotlin/com/hoc081098/flowext/internal/Symbols.kt
@@ -25,9 +25,11 @@
 package com.hoc081098.flowext.internal
 
 import com.hoc081098.flowext.utils.Symbol
+import kotlin.jvm.JvmField
 
 /*
  * Symbol used to indicate that the flow is complete.
  * It should never leak to the outside world.
  */
+@JvmField
 internal val DONE_VALUE = Symbol("DONE_VALUE")